vkontakt/requirementx.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
choicelib==0.1.5
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.30.0
sniffio==1.3.0
typing_extensions==4.5.0
urllib3==2.0.2
vbml==1.1.post1
vk-api==11.9.9
vkbottle==4.3.12
vkbottle-types==5.131.146.15
watchfiles==0.18.1
yarl==1.9.2
